moved animatedPlayerCard
